ihixs completed for the ggf channel
gtests: added tests on as_series.cpp

models/model.cpp: added a feature not to perform the evolution running
if we are already at the scale and order requested

pdf/luminosity.cpp: added the posibility of using one pdf set fixed
from the runcard by name. This is stored in the variable UI.pdf_set and
requested by the user in the runcard through the variable pdf_set. Note
that pdf_set and pdf_provider are mutually exclusive: the user can only
define one of them (else he gets an error). The user has to define one
of them, because the default pdf value is None for both variables now.

tools/user_interface.cpp: added the pdf_set variable, with default
value None.